+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Move Config Classes to Config Package
 - Correct BWG Pitcher Plant Item Models
 - Slight Change to FruitBlockProcessor#finalizeProcessing to be more Efficient
+- Add Config to Enable/Disable Spawning of Oddion and Man O War
 
 # 2.2.5
 - Make Pale Pumpkin Set Compostable